+++ .env.ci
@@ -17,6 +17,8 @@ COOKIE_SECRET=a_cookie_secret_1234
 # SERVER_PORT=3000
 AUTO_ACTIVATE_NEW_USERS=false
 VERIFY_EMAILS=true
+ALLOW_USER_SELF_DELETE=true
+LOST_PASSWORD_ENABLE=true
 
 # OAuth providers
 PROVIDER_SUCCESS_REDIRECT=http://localhost:3000/healthz
@@ -46,4 +48,3 @@ SMTP_USER=user
 SMTP_SECURE=false
 SMTP_SENDER=hbp@hbp.com
 
-ALLOW_USER_SELF_DELETE=true
