+++ .env.example
@@ -5,11 +5,15 @@
 
 ################## General Settings ###################
 ## Required: Hasura GraphQL endpoint
+## Example: http://your-hasura-server.com/v1/graphql
 HASURA_ENDPOINT=
+
 ## Required: Hasura admin secret
 HASURA_GRAPHQL_ADMIN_SECRET=
+
 ## Strongly recommended: public URL for HBP (used in email templates and OAuth callbacks)
 SERVER_URL=
+
 # PORT=3000
 
 ## Automatically loads the SQL schema and Hasura metadata on startup.
@@ -75,12 +79,17 @@ JWT_KEY=
 # REDIRECT_URL_ERROR=
 
 #---------------------- Emails -----------------------#
+## Note: the following settings require valid SMTP configuration
+
 ## List of comma-separated email domains allowed to register
 ## Example: ALLOWED_EMAIL_DOMAINS=gmail.com,hotmail.com
 # ALLOWED_EMAIL_DOMAINS=
 
 ## Allow users to change their email
 # CHANGE_EMAIL_ENABLE=true
+
+## Allow users to reset their passwords from a ticket sent by email
+# LOST_PASSWORD_ENABLE=false
 
 ## Send an email notification to the previous email when changed
 # NOTIFY_EMAIL_CHANGE=false
