+++ src/routes/auth/change-password/index.ts
@@ -1,5 +1,13 @@
 import { Router } from 'express'
-import request from './request'
-import change from './change'
 
-export default Router().post('/request', request).post('/change', change)
+import { LOST_PASSWORD_ENABLE } from '@shared/config'
+
+import lost from './lost'
+import change from './change'
+import reset from './reset'
+
+const router = Router().post('/', change)
+
+if (LOST_PASSWORD_ENABLE) router.post('/lost', lost).post('/reset', reset)
+
+export default router
