+++ src/routes/auth/change-password/lost.ts
@@ -1,13 +1,13 @@
-import { asyncWrapper, selectAccountByEmail } from '@shared/helpers'
 import { Request, Response } from 'express'
+import { v4 as uuidv4 } from 'uuid'
+import Boom from '@hapi/boom'
 
+import { asyncWrapper, selectAccountByEmail } from '@shared/helpers'
 import { SMTP_ENABLE, SERVER_URL } from '@shared/config'
 import { emailClient } from '@shared/email'
 import { forgotSchema } from '@shared/validation'
 import { setNewTicket } from '@shared/queries'
 import { request } from '@shared/request'
-import { v4 as uuidv4 } from 'uuid'
-import Boom from '@hapi/boom'
 
 /**
  * * Creates a new temporary ticket in the account, and optionnaly send the link by email
