+++ src/shared/config/authentication/index.ts
@@ -15,3 +15,4 @@ export const NOTIFY_EMAIL_CHANGE = castBooleanEnv('NOTIFY_EMAIL_CHANGE', false)
 export const ANONYMOUS_USERS_ENABLE = castBooleanEnv('ANONYMOUS_USERS_ENABLE', false)
 export const ALLOW_USER_SELF_DELETE = castBooleanEnv('ALLOW_USER_SELF_DELETE', false)
 export const VERIFY_EMAILS = castBooleanEnv('VERIFY_EMAILS', false)
+export const LOST_PASSWORD_ENABLE = castBooleanEnv('LOST_PASSWORD_ENABLE', false)
